Changed string formatting length parameter for "%.*s" to int, as that's expected by the standard

+++ Oxygen/oxygenengine/source/oxygen/menu/devmode/windows/AudioPlaybackWindow.cpp
@@ -79,7 +79,7 @@ void AudioPlaybackWindow::buildContent()
 
 			ImGui::TableSetColumnIndex(1);
 			const std::string_view audioKey = audioPlayer.getPlayingSoundAudioKey(ref);
-			ImGui::Text("%.*s", audioKey.length(), audioKey.data());
+			ImGui::Text("%.*s", (int)audioKey.length(), audioKey.data());
 
 			if (mShowPosition)
 			{
